rename some package name and add a dao level to access the database
