+++ BaseModel/bin/.gitignore
@@ -1,1 +1,2 @@
+/CSP.jpg
 /com/
