+++ BaseModel/src/com/wpi/cs509/teamA/dao/DupEntranceMapDao.java
@@ -1,0 +1,5 @@
+package com.wpi.cs509.teamA.dao;
+
+public class DupEntranceMapDao {
+
+}
