+++ BaseModel/src/com/wpi/cs509/teamA/dao/impl/DupEntranceMapDaoImpl.java
@@ -1,0 +1,5 @@
+package com.wpi.cs509.teamA.dao.impl;
+
+public class DupEntranceMapDaoImpl {
+
+}
