+++ BaseModel/src/com/wpi/cs509/teamA/findRoute/impl/FindRouteImpl.java
@@ -1,4 +1,4 @@
-package com.wpi.cs509.teamA.findRouteImpl;
+package com.wpi.cs509.teamA.findRoute.impl;
 
 import java.util.List;
 import java.util.Map;
